pyproject.toml
[project]
name = "isochrones"
version = "0.1.0"
description = "Tools for geocoding addresses and calculating isochrones using GraphHopper API"
authors = [{name = "Your Name", email = "your.email@example.com"}]
maintainers = [{name = "Your Name", email = "your.email@example.com"}]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}
dependencies = [
    "duckdb",
    "duckdb-extensions",
    "duckdb-extension-spatial",
    "python-dotenv>=1.0.0",
    "playwright>=1.53.0",
    "superclaude>=3.0.0.1",
    "panel>=1.7.4",
    "pydeck>=0.9.1",
    "geopandas>=1.1.1",
    "pyarrow>=21.0.0",
]



# Note: Dependencies are now primarily defined in each script using inline script metadata
# These are kept for reference and for traditional package installation
[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.7",
]
geocode = [
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
    "duckdb",
    "duckdb-extensions",
    "duckdb-extension-spatial",
    "python-dotenv>=1.0.0",
]
server = [
    "fastapi>=0.104.1",
    "uvicorn>=0.23.2",
    "python-multipart>=0.0.6",
    "duckdb",
    "duckdb-extensions",
    "duckdb-extension-spatial",
    "panel>=1.2.3",
    "param>=2.0.0",
    "pydeck>=0.8.0",
    "holoviews>=1.17.1",
    "bokeh>=3.2.2",
    "python-dotenv>=1.0.0",
]
query = [
    "duckdb",
    "duckdb-extensions",
    "duckdb-extension-spatial",
]
viewer = [
    "duckdb",
    "duckdb-extensions",
    "duckdb-extension-spatial",
    "panel>=1.3.8",
    "param>=2.0.1",
    "pydeck>=0.8.0",
    "python-dotenv>=1.0.0",
]
export-shapefiles = [
    "geopandas>=0.14.0",
    "shapely>=2.0.1",
    "fiona>=1.9.4",
    "pyogrio>=0.7.2",
    "tqdm>=4.66.1",
    "python-dotenv>=1.0.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100
[tool.ruff.lint]
select = ["E", "F", "B", "I", "N", "UP", "S", "BLE", "A", "C4", "T10", "ICN"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"


